+++ CHANGELOG.md
@@ -24,7 +24,9 @@
 * `--fail-fast` option for rspec by default
 * Replace usage of `HashHelper.get_hash_from_array_with_two_parameters`
   with `TestrailApiObject#name_id_pairs`
-* Remove unused `TestrailHelper#ignored_parameters`
+* Remove unused `TestrailHelper#ignored_parameters`,
+  `TestrailHelper#add_result_by_case_name`,
+  `TestrailHelper#mark_rest_environment_dependencies`
 
 ## 0.3.0 (2021-04-05)